Add TrecRun normalize rescore method

+++ tests/test_fusion.py
@@ -31,7 +31,6 @@ class TestSearch(unittest.TestCase):
         os.system(
             f'python -m pyserini.fusion --method rrf --runs {qruns_str} --output {self.output_path} --runtag test')
         self.assertTrue(filecmp.cmp(verify_path, self.output_path))
-        os.remove(self.output_path)
 
     def test_interpolation_fusion_simple(self):
         input_paths = ['tests/resources/simple_trec_run_fusion_1.txt', 'tests/resources/simple_trec_run_fusion_2.txt']
@@ -41,7 +40,6 @@ class TestSearch(unittest.TestCase):
         os.system(
             f'python -m pyserini.fusion --method interpolation --alpha 0.4 --runs {qruns_str} --output {self.output_path} --runtag test')
         self.assertTrue(filecmp.cmp(verify_path, self.output_path))
-        os.remove(self.output_path)
 
     def test_average_fusion_simple(self):
         input_paths = ['tests/resources/simple_trec_run_fusion_1.txt', 'tests/resources/simple_trec_run_fusion_2.txt']
@@ -51,7 +49,6 @@ class TestSearch(unittest.TestCase):
         os.system(
             f'python -m pyserini.fusion --method average --runs {qruns_str} --output {self.output_path} --runtag test')
         self.assertTrue(filecmp.cmp(verify_path, self.output_path))
-        os.remove(self.output_path)
 
     def test_reciprocal_rank_fusion_complex(self):
         os.system('wget -q -nc https://www.dropbox.com/s/duimcackueph2co/anserini.covid-r2.abstract.qq.bm25.txt.gz')
@@ -68,9 +65,11 @@ class TestSearch(unittest.TestCase):
             f'python -m pyserini.fusion --method rrf --runs {qruns_str} --output {self.output_path} --runtag reciprocal_rank_fusion_k=60')
         verify_path = 'anserini.covid-r2.fusion1.txt'
         self.assertTrue(filecmp.cmp(verify_path, self.output_path))
-        os.remove(self.output_path)
         os.system('rm anserini.covid-r2.*')
-
+    
+    def tearDown(self):
+        if os.path.exists(self.output_path):
+            os.remove(self.output_path)
 
 if __name__ == '__main__':
     unittest.main()
